refactor(CalculationController): Controller 및 Service 역할 분리 및 명칭 변경
- CalculationHandler를 CalculationService로 리팩토링하여 역할을 명확하게 구분
- FlowFactory를 통해 CalculationController 객체를 생성하도록 구조 개선
- Controller와 Service 계층 간 책임을 명확히 분리하여 코드 가독성 및 유지보수성 향상
